Move BaseDbContext and app.db to Database folder
