+++ Repositories/DbUserRepository.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
-using Karma.data.messages;
+using Karma.Database;
 using Karma.Models;
 
 namespace Karma.Repositories
