+++ appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "DataSource=app.db"
+    "DefaultConnection": "DataSource=DataBase/app.db"
   },
   "JwtSettings" : {
     "Secret": "Super secret code"
